backend/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<4.0
django-allauth[mfa]>=65.0,<66.0
dj-rest-auth>=7.0,<8.0
djangorestframework-simplejwt>=5.3,<6.0
celery[redis]>=5.4,<6.0
django-celery-beat>=2.7,<3.0
channels>=4.1,<5.0
channels-redis>=4.2,<5.0
daphne>=4.1,<5.0
django-encrypted-model-fields>=0.6,<1.0
django-filter>=24.0,<25.0
django-cors-headers>=4.4,<5.0
argon2-cffi>=23.1,<24.0
psycopg2-binary>=2.9,<3.0
boto3>=1.35,<2.0
anthropic>=0.40,<1.0
twilio>=9.0,<10.0
gunicorn>=22.0,<23.0
whitenoise>=6.7,<7.0
redis>=5.0,<6.0
# Testing
pytest>=8.0,<9.0
pytest-django>=4.9,<5.0
pytest-asyncio>=0.24,<1.0
factory-boy>=3.3,<4.0
pytest-cov>=5.0,<6.0
freezegun>=1.4,<2.0
